Fix initialSpaces imports after folder restructure
