+++ src/config/initialSpaces/initialProfileSpace.ts
@@ -1,7 +1,7 @@
 import { SpaceConfig } from "@/app/(spaces)/Space";
 import { FeedType, FilterType } from "@neynar/nodejs-sdk/build/api";
 import { getLayoutConfig } from "@/common/utils/layoutFormatUtils";
-import { INITIAL_SPACE_CONFIG_EMPTY } from "../../initialSpaceConfig";
+import { INITIAL_SPACE_CONFIG_EMPTY } from "./initialSpaceConfig";
 import { deepClone } from "@/common/lib/utils/deepClone";
 
 // Set default tabNames for profile spaces
