+++ src/config/initialSpaces/initialProposalSpace.ts
@@ -1,5 +1,5 @@
 import { SpaceConfig } from "@/app/(spaces)/Space";
-import { INITIAL_SPACE_CONFIG_EMPTY } from "../../initialSpaceConfig";
+import { INITIAL_SPACE_CONFIG_EMPTY } from "./initialSpaceConfig";
 import { Address } from "viem";
 import { deepClone } from "@/common/lib/utils/deepClone";
